forgive me for this spaghetti
