+++ src/routes/art/genuary/2022/index.js
@@ -1,0 +1,5 @@
+import { day1 } from "./day1"
+
+export default {
+  day1
+}
