+++ static/style.css
@@ -106,6 +106,7 @@ p {
 }
 
 .link,
+h3 a,
 p a, 
 li a,
 blockquote a {
